🚚 rename to createDatabase
